origin

+++ Web-Project/src/main/java/ru/itis/inform/models/FilmExistance.java
@@ -3,16 +3,19 @@ package ru.itis.inform.models;
 /**
  * Created by Тимур on 12.11.2016.
  */
+
+// Модель сущ фильма и его цены
 public class FilmExistance {
     private int id;
     private int quantity;
-    private double money;
+    private int money;
 
 
-    public FilmExistance(int id, int quantity, double money) {
+    public FilmExistance(int id, int quantity, int money) {
         this.id = id;
         this.quantity = quantity;
         this.money = money;
+
     }
 
     public int getId() {
@@ -23,7 +26,7 @@ public class FilmExistance {
         return quantity;
     }
 
-    public double getMoney() {
+    public int getMoney() {
         return money;
     }
 }
